Move TestableObservable to Observable, using static construction.
